Attempt fix on parsing and compiling error
Temporarily using the parsing input from go-eval as a workaround, as
the parser first expects a "package" value to parse (possibly a change
upon reaching go 1.0). Also fixed a problem where if GOBIN and GOARCH
are not set, the compile() function would break. This assumes a 64-bit
architecture if GOARCH is not set. An ideal way to be done later is to
call "go tool" first to check the possible tools, then use the right
one according (in the regex format "[0-9]g" and "[0-9]l").

A recurring problem is that the World.source() method does not work as
intended, where it cannot correctly display the code in question. It is
because of this that compile() will not do anything.