Added item numbers to source code display
For convenience of picking out each of the items (code, declaration, or
package) of the World struct in the source code, number labels for each
are now shown when using "!". Of course, the source code with these
additions cannot be compiled, so it requires a flag to determine when
it is compiling and when it is being viewed by the user.

This is also for a possible future task of removing specific lines.