Fix declaration processing
Exchanged the order of declarations (i.e. var a = 8) and code, as now
declarations can be run like code, but certain code like method calls
cannot be treated like declarations. As such, it is better to let
ParseStmtList be the fall back to contain the superset (code) of input
rather than for ParseDeclList with the subset (declarations).

A note is that if there really is a need for declarations in code, the
user can use the ":" option to forcefully specify its definition,
although this is not preferred in comparison to initializing with ":="
due to possible instability of code (where the compiler might complain
about it not being in use).