Added write source mode
Added a way to change the mode to only write to source by default
rather than having to use ":" to add it persistently. This does not
run the code unless a "run" command is inputted or the mode is switched
back to REPL mode.

Also moved unstable to be an attribute of World, given it close ties
to the code state itself and for convenience of exec_special().