Extended source code removal
Extended the functionalities of the "-" command to also remove specific parts
of the code, with indices delimited by commas. This allows not just popping off
the list but also removal of code in the middle (i.e. if some function call
in the middle causes some unintentional effects). The original functionality
is preserved.

Also updated the help menu to show new commands and functionalities.